JWT token now authenticates user.
I went a little insane, but I finally got the whole authentication thing to work. Albeit it is a little wonky and not somthn I recommend, it's as good as it gets for now. Maybe sometime soon, I'll work on improving it with express-session, but for now I'll be using cookies to store the JWT token.

Next up will be more frontend stuff and Logout.