Temporary fix: due to PaaS having suffixes in PSL, cross-site cookies cannot be used unless a custom domain is used.
And since I am a broke college student, I have devised a solution that can be utilized for the time being.

JWT is now stored in sessionStorage and most components have been modified to handle such. While it is a security risk for the most part exposing the token like such, this is still a temporary fix until I can find a way to use cookies and sessions again.